refactor: Ensure correct type conversion for access, creation, and li…

+++ src/evaluator/builtins/system/filesystem_linux.go
@@ -18,7 +18,7 @@ func getFileOwnership(info os.FileInfo) (uid, gid uint32, ok bool) {
 // getAccessTime extracts access time from file info (Linux-specific)
 func getAccessTime(info os.FileInfo) (int64, bool) {
 	if stat, ok := info.Sys().(*syscall.Stat_t); ok {
-		return stat.Atim.Sec, true
+		return int64(stat.Atim.Sec), true
 	}
 	return 0, false
 }
@@ -27,7 +27,7 @@ func getAccessTime(info os.FileInfo) (int64, bool) {
 // Note: Linux uses change time (Ctim), not birth time
 func getCreationTime(info os.FileInfo) (int64, bool) {
 	if stat, ok := info.Sys().(*syscall.Stat_t); ok {
-		return stat.Ctim.Sec, true
+		return int64(stat.Ctim.Sec), true
 	}
 	return 0, false
 }
@@ -43,7 +43,7 @@ func getFileInode(info os.FileInfo) (uint64, bool) {
 // getFileLinkCount extracts link count from file info (Linux-specific)
 func getFileLinkCount(info os.FileInfo) (uint64, bool) {
 	if stat, ok := info.Sys().(*syscall.Stat_t); ok {
-		return stat.Nlink, true
+		return uint64(stat.Nlink), true
 	}
 	return 0, false
 }
